requirements.txt
anyio==3.6.1
apscheduler==3.9.1
async-generator==1.10
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chromedriver-autoinstaller==0.4.0
cryptography==38.0.1
dynaconf==3.1.11
exceptiongroup==1.0.0rc9
greenlet==1.1.3.post0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
numpy==1.23.3
outcome==1.2.0
pandas==1.5.0
pycparser==2.21
pyopenssl==22.1.0
pysocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot==20.0a4
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
scweet==1.8
selenium==4.2.0
setuptools==65.4.1
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
sqlalchemy==1.4.41
tornado==6.2
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
urllib3-secure-extra==0.1.0
wsproto==1.2.0
